+++ src/test/java/com/wodle/domain/WordTest.java
@@ -7,7 +7,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-import com.wodle.Result;
 import java.util.List;
 import org.junit.jupiter.api.Test;
